+++ unipept/src/tree/lca.rs
@@ -30,7 +30,7 @@ impl LCACalculator {
 }
 
 impl agg::Aggregator for LCACalculator {
-    fn aggregate(&self, taxons: &HashMap<TaxonId, usize>) -> Result<TaxonId, agg::Error> {
+    fn aggregate(&self, taxons: &HashMap<TaxonId, f32>) -> Result<TaxonId, agg::Error> {
         if taxons.len() == 0 { return Err(agg::Error::EmptyInput); }
         let subtree = try!(SubTree::new(self.root, &self.parents, taxons)).collapse(&Add::add);
         Ok(subtree.root)
